backend/requirements.txt
alembic==1.13.2
email-validator==2.2.0
fastapi==0.111.1
httpx==0.27.0
opencv-python-headless==4.10.0.84
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
Pillow==10.4.0
psycopg[binary]==3.2.1
pydantic-settings==2.4.0
pytest==8.3.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
SQLAlchemy==2.0.31
uvicorn[standard]==0.30.3
